requirements.txt
transformers==4.40.1
accelerate==0.29.3
jiwer==3.0.4
scikit-learn==1.4.2
scikit-image==0.23.2
opencv-python==4.9.0.80
fastapi==0.111.0
hf_transfer==0.1.6
tensorboard==2.17.0
numpy==1.26.4
albumentations==1.4.6
paddleocr==2.8.1
paddlepaddle-gpu==2.6.2
uvicorn==0.31.0
einops
hydra-core
hydra_colorlog
apted
fastapi[standard]
uvicorn
lxml==4.9.3
torchmetrics
torchtext
jsonlines
ptflops
tokenizers
pycocotools
faster-coco-eval
shapely
ultralytics==8.0.153
ruamel.yaml
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1